Add support for traitCollection changes to values passed into .border()
This API lied a little bit, because it allowed you to pass in UIColor
values which would immediately get mapped into CGColor to set the color
on the layer's border. Which meant that traitCollection changes wouldn't
trigger updates to that borderColor.

This resolve that by observing the view's traitCollectionDidChange
selector in addition to the border producer. So when the trait
collection userInterfaceStyle changes, we resend the last border
value and any UIColors get mapped into the correct color for the current
UserInterfaceStyle